Put dataclasses and the StockfishException in a separate file (#166)
* put dataclasses and exceptions in a separate file

* update
