+++ stockfish/__init__.py
@@ -1,3 +1,4 @@
-from .models import Stockfish, StockfishException
+from .models import Stockfish
+from .types import StockfishException
 
 __all__ = ["Stockfish", "StockfishException"]
